+++ VotingApplication/Controllers/UserController.cs
@@ -6,8 +6,6 @@ using System.Threading.Tasks;
 using System.Net;
 using System.Net.Mail;
 using VotingApplication.ViewModels;
-using Newtonsoft.Json;
-using System.Linq;
 
 namespace VotingApplication.Controllers
 {
@@ -31,89 +29,7 @@ namespace VotingApplication.Controllers
             ViewData["UserName"] = HttpContext.User.Identity.Name;
             return View("Dashboard/Index"); //Index view
         }
-
-        [HttpGet]
-        public IActionResult ZipCodeMap()
-        {
-            return View("Dashboard/ZipCodeMap"); //Index view
-        }
-
-        private string stateAbbreviation(string state)
-        {
-            switch (state.ToUpper())
-            {
-                case "ALABAMA": return "AL";
-                case "ALASKA": return "AK";
-                case "ARIZONA": return "AZ";
-                case "ARKANSAS": return "AR";
-                case "CALIFORNIA": return "CA";
-                case "COLORADO": return "CO";
-                case "CONNECTICUT": return "CT";
-                case "DELAWARE": return "DE";
-                case "FLORIDA": return "FL";
-                case "GEORGIA": return "GA";
-                case "HAWAII": return "HI";
-                case "IDAHO": return "ID";
-                case "ILLINOIS": return "IL";
-                case "INDIANA": return "IN";
-                case "IOWA": return "IA";
-                case "KANSAS": return "KS";
-                case "KENTUCKY": return "KY";
-                case "LOUISIANA": return "LA";
-                case "MAINE": return "ME";
-                case "MARYLAND": return "MD";
-                case "MASSACHUSETTS": return "MA";
-                case "MICHIGAN": return "MI";
-                case "MINNESOTA": return "MN";
-                case "MISSISSIPPI": return "MS";
-                case "MISSOURI": return "MO";
-                case "MONTANA": return "MT";
-                case "NEBRASKA": return "NE";
-                case "NEVADA": return "NV";
-                case "NEW HAMPSHIRE": return "NH";
-                case "NEW JERSEY": return "NJ";
-                case "NEW MEXICO": return "NM";
-                case "NEW YORK": return "NY";
-                case "NORTH CAROLINA": return "NC";
-                case "NORTH DAKOTA": return "ND";
-                case "OHIO": return "OH";
-                case "OKLAHOMA": return "OK";
-                case "OREGON": return "OR";
-                case "PENNSYLVANIA": return "PA";
-                case "RHODE ISLAND": return "RI";
-                case "SOUTH CAROLINA": return "SC";
-                case "SOUTH DAKOTA": return "SD";
-                case "TENNESSEE": return "TN";
-                case "TEXAS": return "TX";
-                case "UTAH": return "UT";
-                case "VERMONT": return "VT";
-                case "VIRGINIA": return "VA";
-                case "WASHINGTON": return "WA";
-                case "WEST VIRGINIA": return "WV";
-                case "WISCONSIN": return "WI";
-                case "WYOMING": return "WY";
-                case "GUAM": return "GU";
-                case "PUERTO RICO": return "PR";
-                case "VIRGIN ISLANDS": return "VI";
-                default: return "IA";
-            }
-        }
-        [HttpGet]
-        public IActionResult RequestZipCodes(string state = null)
-        {
-            ZipCodeFeatureCollection collection = new ZipCodeFeatureCollection();
-            foreach (ZipCodeDataModel zipCode in _Context.ZipCode.Where(zip => state != null && zip.State == stateAbbreviation(state)))
-            {
-                ZipCodeFeature feature = new ZipCodeFeature();
-                feature.properties = new Properties(zipCode);
-                feature.geometry = new Geometry(zipCode);
-                collection.features.Add(feature);
-            }
-            ZipCodeFeatureCollectionViewModel jsonData = new ZipCodeFeatureCollectionViewModel();
-            jsonData.ZipCodes = collection;
-            return Content(JsonConvert.SerializeObject(collection), "application/json");
-        }
-
+        
         [HttpGet]
         public IActionResult Profile()
         {
