+++ VotingApplication/Views/Admin/Dashboard.cshtml
@@ -5,3 +5,6 @@
     The dashboard will show you important information in the future.
 </p>
 <a asp-area="" asp-controller="Admin" asp-action="UserManagement" class="nav-link">Manage Users</a>
+<a asp-area="" asp-controller="Admin" asp-action="AddCandidate" class="nav-link">Add Candidate</a>
+<a asp-area="" asp-controller="Admin" asp-action="AddOffice" class="nav-link">Add Office</a>
+<a asp-area="" asp-controller="Admin" asp-action="ZipCodeMap" class="nav-link">View Zip Code Map</a>
