+++ VotingApplication/Views/Home/Index.cshtml
@@ -8,5 +8,4 @@
     Any ballots you are able to vote on.
     Any past ballots that are now finalized.
 </p>
-<a asp-area="" asp-controller="Admin" asp-action="AddCandidate" class="nav-link">Add Candidate</a>
-<a asp-area="" asp-controller="Admin" asp-action="AddOffice" class="nav-link">Add Office</a>
+
