+++ VotingApplication/Views/User/Dashboard/Index.cshtml
@@ -7,8 +7,9 @@
 <a asp-area="" asp-controller="Voter" asp-action="Registration" class="nav-link">Register to Vote</a>
 
 <a asp-area="" asp-controller="Voter" asp-action="Demographics" class="nav-link">Enter Demographic info</a>
-
-<a asp-area="" asp-controller="User" asp-action="ZipCodeMap" class="nav-link">View Zip Code Map</a>
-
+<br/>
 <a asp-area="" asp-controller="Admin" asp-action="Dashboard" class="nav-link">Admin Dashboard</a>
+<label>
+    This link will be removed and Admins will automatically be redirected to the dashboard below.
+</label>
